Move NewRequestID function from handler package to common package
Function can be reused by other packages
