+++ common/identity_test.go
@@ -173,3 +173,16 @@ func TestAddAuthHeaders(t *testing.T) {
 		})
 	})
 }
+
+func TestNewRequestID(t *testing.T) {
+	Convey("create a requestID with length of 12", t, func() {
+		requestID := NewRequestID(12)
+		So(len(requestID), ShouldEqual, 12)
+
+		Convey("create a second requestID with length of 12", func() {
+			secondRequestID := NewRequestID(12)
+			So(len(secondRequestID), ShouldEqual, 12)
+			So(secondRequestID, ShouldNotEqual, requestID)
+		})
+	})
+}
